// src/app/interfaces/interfaces.ts
export interface RespuestaPosts {
  ok: boolean;
  pagina: number;
  posts: Post[];
}

export interface Post {
  imgs?: string[];
  _id?: string;
  mensaje?: string;
  coords?: string;
  usuario?: Usuario;
  created?: string;
}

export interface Usuario {
  avatar?: string;
  _id?: string;
  nombre?: string;
  email?: string;
  password?: string;
}

export interface Paciente{
  _id?: string;
  nombre?: string;
  apellido?:string;
  edad?:number;
  sexo?:string;
  documento?:number;
  fechaNacido?:Date;
  activo?:boolean
}